Base Test added , refactoring , AndroidJunit4 ext dependency added to gradle
