+++ app/src/androidTest/java/com/andor/navigate/notepad/AddNoteTest.kt
@@ -1,32 +1,14 @@
 package com.andor.navigate.notepad
 
 
-import android.view.View
-import android.view.ViewGroup
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.*
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.RootMatchers.withDecorView
 import androidx.test.espresso.matcher.ViewMatchers.*
-import androidx.test.filters.LargeTest
-import androidx.test.rule.ActivityTestRule
-import androidx.test.runner.AndroidJUnit4
-import com.andor.navigate.notepad.auth.AuthActivity
-import org.hamcrest.Description
-import org.hamcrest.Matcher
-import org.hamcrest.Matchers.*
-import org.hamcrest.TypeSafeMatcher
-import org.junit.Rule
+import com.andor.navigate.notepad.base.BaseTest
+import org.hamcrest.Matchers.allOf
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@LargeTest
-@RunWith(AndroidJUnit4::class)
-class AddNoteTest {
-
-    @Rule
-    @JvmField
-    var mActivityTestRule = ActivityTestRule(AuthActivity::class.java)
+class AddNoteTest : BaseTest() {
 
     /*
     * recorder test for reference of creating a Note
@@ -196,23 +178,5 @@ class AddNoteTest {
         )
 
         cancelAddNoteAction.perform(click())
-    }
-
-    private fun childAtPosition(
-        parentMatcher: Matcher<View>, position: Int
-    ): Matcher<View> {
-
-        return object : TypeSafeMatcher<View>() {
-            override fun describeTo(description: Description) {
-                description.appendText("Child at position $position in parent ")
-                parentMatcher.describeTo(description)
-            }
-
-            public override fun matchesSafely(view: View): Boolean {
-                val parent = view.parent
-                return parent is ViewGroup && parentMatcher.matches(parent)
-                        && view == parent.getChildAt(position)
-            }
-        }
     }
 }
