+++ app/src/androidTest/java/com/andor/navigate/notepad/DeleteNoteTest.kt
@@ -1,32 +1,17 @@
 package com.andor.navigate.notepad
 
 
-import android.view.View
-import android.view.ViewGroup
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.*
+import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.action.ViewActions.longClick
 import androidx.test.espresso.contrib.RecyclerViewActions
 import androidx.test.espresso.matcher.ViewMatchers.*
-import androidx.test.filters.LargeTest
-import androidx.test.rule.ActivityTestRule
-import androidx.test.runner.AndroidJUnit4
+import com.andor.navigate.notepad.base.BaseTest
 import com.andor.navigate.notepad.listing.adapter.ListingAdapter
-import org.hamcrest.Description
-import org.hamcrest.Matcher
-import org.hamcrest.Matchers.`is`
 import org.hamcrest.Matchers.allOf
-import org.hamcrest.TypeSafeMatcher
-import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@LargeTest
-@RunWith(AndroidJUnit4::class)
-class DeleteNoteTest {
-
-    @Rule
-    @JvmField
-    var mActivityTestRule = ActivityTestRule(SplashScreenActivity::class.java)
+class DeleteNoteTest : BaseTest() {
 
     @Test
     fun deleteNote() {
@@ -53,23 +38,5 @@ class DeleteNoteTest {
             )
         )
         actionMenuItemView.perform(click())
-    }
-
-    private fun childAtPosition(
-        parentMatcher: Matcher<View>, position: Int
-    ): Matcher<View> {
-
-        return object : TypeSafeMatcher<View>() {
-            override fun describeTo(description: Description) {
-                description.appendText("Child at position $position in parent ")
-                parentMatcher.describeTo(description)
-            }
-
-            public override fun matchesSafely(view: View): Boolean {
-                val parent = view.parent
-                return parent is ViewGroup && parentMatcher.matches(parent)
-                        && view == parent.getChildAt(position)
-            }
-        }
     }
 }
