+++ app/src/androidTest/java/com/andor/navigate/notepad/ListTypeTest.kt
@@ -1,29 +1,13 @@
 package com.andor.navigate.notepad
 
-import android.view.View
-import android.view.ViewGroup
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.matcher.ViewMatchers
-import androidx.test.filters.LargeTest
-import androidx.test.rule.ActivityTestRule
-import androidx.test.runner.AndroidJUnit4
-import com.andor.navigate.notepad.auth.AuthActivity
-import org.hamcrest.Description
-import org.hamcrest.Matcher
+import com.andor.navigate.notepad.base.BaseTest
 import org.hamcrest.Matchers
-import org.hamcrest.TypeSafeMatcher
-import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@LargeTest
-@RunWith(AndroidJUnit4::class)
-class ListTypeTest {
-
-    @Rule
-    @JvmField
-    var mActivityTestRule = ActivityTestRule(AuthActivity::class.java)
+class ListTypeTest : BaseTest() {
 
     @Test
     fun selectGridListAdapter() {
@@ -65,25 +49,6 @@ class ListTypeTest {
             )
         )
         selectGridAction.perform(ViewActions.click())
-    }
-
-
-    private fun childAtPosition(
-        parentMatcher: Matcher<View>, position: Int
-    ): Matcher<View> {
-
-        return object : TypeSafeMatcher<View>() {
-            override fun describeTo(description: Description) {
-                description.appendText("Child at position $position in parent ")
-                parentMatcher.describeTo(description)
-            }
-
-            public override fun matchesSafely(view: View): Boolean {
-                val parent = view.parent
-                return parent is ViewGroup && parentMatcher.matches(parent)
-                        && view == parent.getChildAt(position)
-            }
-        }
     }
 }
